% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{lrm.estimate}
\alias{lrm.estimate}
\title{LRM_estimate}
\usage{
lrm.estimate(lr.model)
}
\arguments{
\item{lr.model}{fitted linear regression using lrm}
}
\value{
estimated parameters and standard error
}
\description{
Estimate the parameters of a linear regression model
}
\examples{
library(Rcpp)
sourceCpp("~/lrm/Cplusplus/lrm_Cpp.cpp")
m = lrm(mtcars$mpg, mtcars$wt)
lrm.estimate(m)

}
