% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{lrm.Ftest}
\alias{lrm.Ftest}
\title{LRM_Ftest}
\usage{
lrm.Ftest(lr.model, alpha = 0.05, interpretation = TRUE, anova = TRUE)
}
\arguments{
\item{lr.model}{the fitted linear regression model}

\item{alpha}{the significance level}

\item{interpretation}{print the result of hypothesis testing}

\item{anova}{print anova table}
}
\value{
hypothesis testing result
}
\description{
Test the sigificance of parameters of a fitted linear regression model
}
\examples{
library(Rcpp)
sourceCpp("~/lrm/Cplusplus/lrm_Cpp.cpp")
m = lrm(mtcars$mpg, mtcars$wt)
lrm.Ftest(m)

}
