% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{lrm.partialtest}
\alias{lrm.partialtest}
\title{LRM_partialtest}
\usage{
lrm.partialtest(
  lr.model,
  test.variable,
  alpha = 0.05,
  confidence.interval = TRUE,
  interpretation = TRUE
)
}
\arguments{
\item{lr.model}{the fitted linear regression model}

\item{test.variable}{the chosen parameter to be tested}

\item{alpha}{the significance level}

\item{interpretation}{print the result of hypothesis testing}

\item{anova}{print anova table}
}
\value{
hypothesis testing result
}
\description{
Test the significance of one parameter of a fitted linear regression model
}
\examples{
m = lrm(mtcars$mpg, mtcars$wt)
lrm.partialtest(m, 2)

}
