% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{lrm}
\alias{lrm}
\title{LRM}
\usage{
lrm(y, x, intercept = TRUE)
}
\arguments{
\item{y}{the dependent variable}

\item{x}{the independent invariable}

\item{intercept}{model include the intercept term or not}
}
\value{
fitted model
}
\description{
Fit a linear regression model to inference and predict
}
\examples{
lrm(mtcars$mpg, mtcars$wt)

}
