% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM.R
\name{lrm.GLH}
\alias{lrm.GLH}
\title{LRM_GLH}
\usage{
lrm.GLH(lr.model, test.matrix, interpretation = TRUE)
}
\arguments{
\item{lr.model}{the fitted linear regression model}

\item{test.matrix}{contrast matrix}

\item{interpretation}{print the result of hypothesis testing}
}
\value{
hypothesis testing result
}
\description{
Global hypothesis test of a subset of parameters in a fitted linear regression model
}
\examples{
m = lrm(mtcars$mpg, mtcars$wt)
T = as.matrix(c(1,-1))
lrm.GLH(m)

}
